Refactor app to allow for better testing
Currently, the app is instantiated as a global variable and imported by each module. This limits testing as it does not allow for the setup of different app configurations. This is solved by moving app initialisation into an app factory function that can be called as needed.

In place of a single app instance, separate packages are created with Flask Blueprints to encapsulate main app functions into components. Two such blueprints are registered: main and auth. As the app grows, new components will be added for related functions.