Refactor test pass: edit article
- replace insert and update statements with single upsert statement

In the first pass of this test, I tried to use upsert but it failed.
I now realise it failed because I tried to define an attribute of a
possibly NoneType object (article.id) to plug into the
query constructor. This caused the route to fail before the upsert
query could be constructed. Here, I plug in a possibly NoneType
object (article_id), taken from URL parameters, to the constructor
directly.