Separate article creation and edition logic into separate routes
During the first iteration of the controller, I made the decision
to handle the logic of article creation and edition in the same route
to avoid code repetition. However, the need to differentiate the
modes within the route has added complexity and repetition of
conditionals. While it was fun to play with upsert statements,
it seems that dividing the routes would make my code cleaner.

- refactor create_article route
    - instantiate model objects before addition to the session
    - flush article object to allow access to article ID for use in remaining inserts
    - render html with form only
- create edit_article route
    - replace upsert statements with update statements
    - render html with form and article data

+ query database session object rather than database objects directly
    - according to syntax in SQLAlchemy docs