Add test: add and update article categories
- register and sign user in
- post article with category data
- get article with category data
- post article with updated category data
- get article with updated category data

I was stuck for a long time on the issue of posting FieldForm values.
I had tried sending the data as a list (as it prints from the route),
I had also tried sending the data as a dict (as advised on SO).
In the end, the issue was that I was using the dict() constructor,
which does not allow hyphens in key names. By writing my own dict,
I was able to use the key names set by WTForms (*-1, *-2, ...)