Add author or admin authentication to article actions
- add function decorator to authenticate current_user is article author or in admin list
- decorate edit_article, publish_article, and delete_article routes
- delete repeated author authentication code in each route

I originally set the decorator to check if the current_user was in a list of roles.
These roles were generated from a database query of the Roles model which had
a many-to-many relationship to the User model, along with association table.
To pass a role as an argument to the decorator, the decorator itself has to be wrapped
in a decorator factory. If I am to devolve publishing powers to other users,
and thus require differentiation between admin and publisher, this is what I would do.