Fix bug: delete button covered by preceeding element
In order to tighten the whitespace between paragraph content inputs,
I added a negative margin to the following element.
However, I realised that this covers the lower half of the preceding
delete button. By giving the delete buttons a relative position,
it places the element on top. If needed, I could further apply a z-index,
but it doesn't seem necessary.

Note, I would also like to increase the whitespace above the latter
labels but ideally I would need this to occur ONLY when preceded
by the .content-placeholder button, not the content input.
I don't know how to select for that.