Remove ORM relationship on Paragraph to Summary
!! Better solution required !!

This relationship is causing me problems that I cannot yet solve.
Rather than continue to burn time, I will continue with this solution
even though I feel it simply avoids the problem.

I want to copy an article along with its paragraphs and summaries.
To that end, I query the orinal Article object and create a new Article
instance with all of the same properties. I then loop through the
collections of the original Article, create new instances of the related
objects, and append them to the collections of the newly copied Article.

This works well until I copy summaries. A summary has a foreign key to
its paragraph and its article, along with ORM mapping. Despite my
efforts, I cannot break the relationship from paragraph to summary
between articles. Both the original and copied paragraphs have
relationships to both the original and copied summaries. A change to
one affects the other. This is unaccceptable.

I have tried to:
  - expunge and make_transient the original and copied objects
  - select the original object via the session and ORM
  - add the copied object via the session and ORM
      - appending to article and paragraph collections
  - flushing and committing between stages
  - clearing collections before and after copying

I have tried brute-forcing combinations thereof. So far, the only thing
to work is to remove the relationship completely, but this feels to me
like curing the disease by killing the patient.