Add URL slugs to active published articles
Display only active published articles with optional URLs slugs.
Publisher deleted articles are inactive unless they are transferred
to a new publisher.

Rather than setting an article's URL to a slug of its title, which
would indirectly impose unique title constraints, it is set to the
unique ID of its publishing note. If the article is requested with
the slug misspelt or omitted, the page is reloaded with the slug
restored.

A method is added to the PublishingNote model to Slugify the article's
title and store it as an object attribute. It is called upon publication
and republication so that it always matches the current title.

Article activity is controlled upon publication, deletion, and
transference. Deleted articles are set as inactive and taken offline.
Requests for deleted articles are redirected to the index page with
user feedback.

A template is rendered for view_artilce but requires development. This
template is copied from, and will be copied back to, preview_article.
It is accessible from the published tab of author's articles and the
active tab of publisher's articles.

The form and script which alert an article's hyperlink upon click are
updated with the PublishingNote ID and slug. A better alternative to
alert needs to be researched.

Documentation is added and amended where relevant.

Models
- edit PublishingNote

Routes
- edit
    - publish_article
    - view_article

Templates
- add view-Article
- edit
    - author-articles
    - publisher-articles

Scripts
- edit display-articles