Refactor: generate image URLs dynamically
Save images to the database without a relative file path.
Instead, create a new route to serve as an endpoint to return the
image file path, and use the url_for function to point to it.

This refactor is necessary after the inclusion of a custom converter
for the view_article route. When building the URL for an image, Flask
prepends the converter to the image URL, causing a 404.

Flask provides a solution in the form of url_for('static', filename='')
which allows URLs to be generated without custom converters. However,
it only points to the static directory, not its subdirectories, and
therefore requires a dedicated endpoint: return_image.

The new route is added and relevant templates updated. The image
upload path constant in config is updated to include the app path that
was removed from add_image route.

App
- edit config

Routes
- add return_image

Templates
- edit
    - admin-articles
    - publisher-articles
    - author-articles
    - publisher-requests
    - edit-article
    - preview-article
    - view-article