Fix: conflated model rel. upon article update
Off the back of the discovery of ORM behaviour in the previous commit,
it is now clear that association objects cannot simply be copied from a
collection as their relationships are also copied and thus conflated.

Instead, new objects must be instantiated and serialised, and then
appended to the parent object.

This doesn't seem to occur with association tables.