Feat: build out article view and preview
Render an article's main body content using a custom web component,
into which encapsulate the styles and methods related to summary
paragraphs.

To control the display of selected reading levels, push a paragraph's
summaries into an array that is set as a property of the custom
webcomponent. Upon selection to increase or decrease the reading level,
iterate through this array and display the next level. Should the
beginning or end of the array be reached, remove the control to further
decrease or increase the level, respectively.

Add additional functionality to allow the user to set a default reading
level, change the font size, and print or save the article with or
without images.