Feat: link article categories to filter page
Allow users to click on a category from an article view to see a list
of articles in the same category.

Append each category in an article view page with a link to the article
filter page with a URL parameter for that category.

On the article filter page, parse the URL parameter and filter the
articles accordingly.