Refactor: reorganise environment variables
Set Flask configuration variables in .flaskenv and commit the file to
the repository. This allows the use of the 'flask' command.

Set application configuration variables (secret key, mail config) in
.env using python-dotenv library. This is ignored by .gitignore and
will not be committed to the repository as it MUST be kept private.

Reconfigure the app config file to load the .env file from the base
directory.