Add: website footer to be displayed sitewide
Add a footer to the layout template so that it is displayed across
the entire site.
Add styling so that the footer is always displayed at the bottom of
the page regardless of the height of the main content.
Update the print media query so that the footer is not printed when
printing an article.