Add: header navigation web accessibility
Replace the header nav icon with a button in order to allow keyboard
access to toggle the menu display. Update its class and style so that
it still looks like the menu icon.
Add appropriate aria labels. Update the script that opens/closes the
menu so that it also toggles these labels.