Feat: custom image validation for alt and citation
Alert users if they upload an image without an image alt (description)
or citation.
Prevent form submission while any outstanding alert exists.

This was a tricky one. I couldn't use WTForm validation as my image
components are dependent on file submission. If a form fails, the file
is reset (browser security measure), which results in the component
being reset. This leaves no space to append the error message to the
relevant input. Moreover, the user would lose all inputted image data
(image, alt, and citation) across the entire form.

Instead, I have created my own validation. Upon expanding the image
component, the alt and citation inputs are rendered with an alert
message. The alert message is removed if the user inputs data and
re-appended if the input is cleared. If the user tries to submit the
form while any alert messages remain, form submission is prevented and
the alert is scrolled into view.

This way, users are forced to add an alt and citation to all images
while retaining all image data that they have already entered.