Feat: set up database migration with flask-migrate
Following instruction from the Miguel Grinberg tutorial:
https://blog.miguelgrinberg.com/post/the-flask-mega-tutorial-part-iv-database

Drop the current database (start fresh).
Create the migration repository.
Generate an initial migration script.
Apply the changes to development database.
Commit

With this extension, the database no longer needs to be dropped and
re-added in order to make changes to the schema. This means I'll be
able to keep any content that has been added.

Moving forward, the development workflow is as follows:
- make desired changes to database models and app
- run 'flask db migrate' to generate the migration script
    - -m "" flag can be used to add a description
- review the generated script
- run 'flask db upgrade' to apply the migration script
- test the changes in the development environment
- (optional) run 'flask db downgrade' to remove changes
    - unwanted scripts can be deleted
- commit the changes with the migration script

To deploy the changes, push the new version of the app and run
'flask db upgrade' on the production server.