pgsql: ensure join queries are declared in order
Where queries to the database contain multiple tables, ensure that the
tables are joined in order.

It seems as though this is another quirk of PostgreSQL compared to
SQLite. In this example, I had joined User to Publisher before either
User or Publisher had themselves been joined to the query. Simply
reordering the joins has solved the issue.