pgsql: replace composite PKs with UniqueConstraint
Replace the composite primary keys of Paragraph and Summary with a
dedicated primary key column. Use the UniqueConstraint table-level
construct to enforce the uniqueness of the related columns.

Remove the prior (ineffective) solution to declare the bidirectional
relationship on Paragraph to Summary.

This change is in response to the ProgrammingError thrown by PostgreSQL.
I provide full details on Stack Overflow (see link below) and await a
more direct solution.

https://stackoverflow.com/questions/71056033/sqlalchemy-invalid-foreign-key-with-postgresql-but-not-sqlite