pgsql: update datetime string formatting
Update the database query for the /filter-articles view to
usefunc.to_char (PostgreSQL) instead of func.strftime (SQLite).
func.strftime is still used to manipulate the string returned from the
database for article data-time attributes in the template.

Note, PostgreSQL requires that the database be queried with the function
applied to the select statement and then grouped with the same function,
rather than applying the function to the group_by clause only. 'FM' is
prepended to the string format to remove padding. Finally, for the
select element options in the template, bracket notation is used to
access the string from the returned dict.