Feat: prompt users to register while on about page
Display a modal box to prompt users to create an account in order to
write their own articles.
For mobile devices, where the navigation menu is collapsed, condition
the prompt box to display only when the menu is collapsed.
Display the prompt only from the /about page to reduce clutter.

I've added this prompt mostly so that prospective employers are aware
that this functionality exists. I want them to see that I've created a
custom CMS.