Test: image upload to s3, update for src and cite
Re-add and commit test images to the static folder in order to test
image uploads.
Test the recently added inputs for image citations.
Update /add-image route to return image src in JSON response.
Update assertions to check image src from JSON, not database.

Note, no changes to the test suite are required to check that images
are uploaded to Amazon s3; if the upload failed, the current assertions
that the image is present on the page would fail.