Fix: summary arrow controls display on Safari
In the summary display web component, reset the margin and padding on
the parent div to zero.

This fix is required because, on Safari browsers, the summary control
arrows were displayed further inside the summary than on Chrome. I
think this is due to Safari not offering the same level of support for
web components; the base reset that I declared in general.css was not
applied to the encapsulated styling within my web components.