Fix: image upload on mobile
Add .jpeg to the accepted file types in Config and form validation.
Convert all JPEGs to .jpeg so as to match imghdr return values.
Update the image validation function in publish/routes so that it does
not replace mime types.

This one took me a while to figure out. I discovered it when uploading
a .jpeg from mobile. However, adding .jpeg to the allowed file types
did not solve the problem. It turns out that the root issue was due to
the use of the imghdr library in the image validation function; imghdr
returns all JPEGs as .jpeg. The image validation function that I pulled
from Miguel Grinberg's tutorial compensated for this by returning .jpeg
as .jpg but this led to mismatched mime types for .jpeg. To solve this,
I convert all JPEGs to .jpeg before image validation.