Feat: alert users when uploaded image is too large
Pop up an alert when a user tries to upload an image that is larger
than 1MB.
Differentiate the alert message for files that are too large against
the catch-all 'invalid' message.

In order to extract the HTTP content (to get the 413 status code), as
well as the JSON body content (to get the image ID), nest the second
promise within the first. The resulting DOM manipulation can then be
placed within the second promise and conditioned upon a successful
upload.

This differs from the way I had originally set up the fetch method; I
had incorrectly assumed that a failed HTTP status would trigger the
catch handler and so placed the failed upload DOM manipulation there.
It worked because the JSON body could not be read, which then triggered
the catch handler.