Style: update see/hide password icon for autofill
Update the styling of the see/hide password icon in auth forms to
increase its contrast for browser autofilled inputs.

Remove the background colour.
Select a darker colour for the icon.

This change is required so as to make the icon move visible when
browsers add a background colour on inputs that are autofilled. I don't
want to remove the autofilled colour as it goes against expected
behaviour.