Feat: delete uploaded but unused images
Delete images that are uploaded but not used in articles from the
database and Amazon S3 cloud storage using APScheduler.

Install, instantiate and initialise the APScheduler extension, and use
the BackgroundScheduler function configured to use the database to
persist scheduled jobs.
Add a helper function, delete_unused_images, to the publish module to
delete Image model objects from the database and to delete image files
from cloud storage. Condition deletion upon the image not being marked
as 'used'. Remove the scheduled job after execution.
4 hours after image upload, enough time for users to complete the
article form, trigger the helper function using the scheduler.
Add or update the effected function docstrings.

This feature is added to avoid filling the database and cloud storage
with unused content. It is required due to the asynchronous nature of
image upload; if an image were uploaded but the article form cancelled,
the image/database object would remain.

The scheduler makes use of the database to persist jobs in case Heroku's
dynos are cycled (reset) before a scheduled job is complete. The
BackgroundScheduler is used so as not to prevent users from completing
article creation/edition.