Fix: display articles in order
Return articles in reverse order so that the newest articles are
displayed first.
For published articles, order the articles by publishing date.
For /display-author-articles, which includes unpublished articles,
additionally order the articles by article ID.

Note, no change to the display order has been made for publishing
requests. This is so that the oldest requests will be seen first.