Fix: allow APScheduler to run in debug mode
Start the APScheduler outside the condition that the app is not running
in debug mode.

This change is required since adding the flask_talisman extension, which
forces all http requests to https; running flask on local host fails
unless it is in debug mode. This worked fine but stopped the scheduler
starting as it was conditioned upon the app not being in debug mode.